requirements.txt
gradio
google-generativeai>=0.7.2
gitpython